Update README.md
Write a program that:

Declares a variable suitable for holding a person’s name
Prompts the user to enter a person’s name with the text Enter name:
Reads the user's input and stores it in the variable created in step 1
Outputs the following information about the name they entered
a. The index of the first character of the last name
b. The first 3 characters of the first name
c. The last 3 characters of the last name